Add npm workspace support, detect local packages from workspace definitions
