+++ src/deploy/functions/packaging/adaptPackageJson.ts
@@ -1,0 +1,6 @@
+import { PackageJson, WorkspacePackageRegistry } from "./findWorkspacePackages";
+import { readTypedJson } from "./utils";
+
+export function adaptPackageJson(packageJsonPath: string, registry: WorkspacePackageRegistry) {
+  const packageJson = readTypedJson<PackageJson>(packageJsonPath);
+}
